+++ ArrayHash/arrayhash_perl.c
@@ -41,7 +41,7 @@ encode_get_type(void *data)
 }
 
 KHARON_DECL char *
-encode_get_scalar(void *data, int *len)
+encode_get_scalar(void *data, size_t *len)
 {
 	SV	*in = data;
 	char	*ret;
